+++ react-complete-guide/src/components/Button.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Button = props => (
+  <button className={props.className || ''} onClick={props.onClick}>
+    {props.text}
+  </button>
+);
+
+export default Button;